fix: database bug

+++ app/src/main/java/com/example/malika/Item.kt
@@ -8,7 +8,7 @@ import java.math.BigDecimal
 @Entity(tableName = "cart_table")
 data class Item constructor(
     @ColumnInfo(name = "name") val name: String?,
-    @ColumnInfo(name = "price") val price: BigDecimal?,
+    @ColumnInfo(name = "price") val price: Int?,
     @ColumnInfo(name = "amount") val amount: Int?
 ) {
     @PrimaryKey(autoGenerate = true) var id:Int = 0
